Move old status EventFilter to a utils file
